+++ QuickfileOptions.cs
@@ -5,6 +5,7 @@ public class QuickfileOptions
     public string AccountNumber { get; set; } = string.Empty;
     public string ApiKey { get; set; } = string.Empty;
     public string ApplicationId { get; set; } = string.Empty;
+    public string WebhookSecret { get; set; } = string.Empty;
     public QuickfileFormat Format { get; set; } = QuickfileFormat.Json;
 }