+++ ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
+using Quickfile.Net.Webhook;
 
 namespace Quickfile.Net;
 
@@ -8,11 +9,11 @@ public static class ServiceCollectionExtensions
     public static IServiceCollection AddQuickfile(this IServiceCollection services, Action<QuickfileOptions> configureOptions)
     {
         services.Configure(configureOptions);
-        services.AddHttpClient<QuickfileClient>((sp, client) =>
-        {
-            var options = sp.GetRequiredService<IOptions<QuickfileOptions>>().Value;
-            // Additional HttpClient configuration can be added here if needed
-        });
+        
+        services.AddSingleton(sp => sp.GetRequiredService<IOptions<QuickfileOptions>>().Value);
+
+        services.AddHttpClient<QuickfileClient>();
+        services.AddTransient<QuickfileWebhookParser>();
 
         return services;
     }